+++ verifypublishsize.bash
@@ -6,8 +6,8 @@ npm pack .
 size="$(wc -c <*.tgz)"
 echo "size is $size"
 
-if [ "$size" -gt 150000 ]
+if [ "$size" -gt 200000 ]
 then
-	echo "Size exceeds 150kB, abort publish"
+	echo "Size exceeds 200kB, abort publish"
 	exit 1
 fi